+++ packages/devtools_app/lib/devtools_app.dart
@@ -22,6 +22,8 @@ export 'src/screens/debugger/span_parser.dart';
 export 'src/screens/debugger/syntax_highlighter.dart';
 export 'src/screens/deep_link_validation/deep_links_controller.dart';
 export 'src/screens/deep_link_validation/deep_links_screen.dart';
+export 'src/screens/dtd/dtd_tools_controller.dart';
+export 'src/screens/dtd/dtd_tools_screen.dart';
 export 'src/screens/inspector/inspector_controller.dart';
 export 'src/screens/inspector/inspector_screen_body.dart';
 export 'src/screens/inspector/inspector_tree_controller.dart';
@@ -55,6 +57,7 @@ export 'src/screens/vm_developer/object_inspector/class_hierarchy_explorer.dart'
 export 'src/screens/vm_developer/object_inspector/class_hierarchy_explorer_controller.dart';
 export 'src/screens/vm_developer/object_inspector/object_inspector_view_controller.dart';
 export 'src/screens/vm_developer/object_inspector/vm_object_model.dart';
+export 'src/screens/vm_developer/queued_microtasks/queued_microtasks_controller.dart';
 export 'src/screens/vm_developer/vm_developer_tools_controller.dart';
 export 'src/screens/vm_developer/vm_developer_tools_screen.dart';
 export 'src/screens/vm_developer/vm_service_private_extensions.dart';