+++ packages/devtools_test/lib/src/mocks/fake_vm_service_wrapper.dart
@@ -17,6 +17,7 @@ class FakeVmServiceWrapper extends Fake implements VmServiceWrapper {
   FakeVmServiceWrapper(
     this._vmFlagManager,
     this._timelineData,
+    this._queuedMicrotasks,
     this._socketProfile,
     this._httpProfile,
     this._memoryData,
@@ -74,6 +75,7 @@ class FakeVmServiceWrapper extends Fake implements VmServiceWrapper {
 
   final VmFlagManager _vmFlagManager;
   final PerfettoTimeline? _timelineData;
+  final QueuedMicrotasks? _queuedMicrotasks;
   SocketProfile? _socketProfile;
   final List<SocketStatistic> _startingSockets;
   HttpProfile? _httpProfile;
@@ -373,6 +375,17 @@ class FakeVmServiceWrapper extends Fake implements VmServiceWrapper {
   Future<Success> clearVMTimeline() => Future.value(Success());
 
   @override
+  Future<QueuedMicrotasks> getQueuedMicrotasks(String isolateId) {
+    if (_queuedMicrotasks == null) {
+      throw StateError(
+        'An argument to the queuedMicrotasks parameter was not provided to '
+        'createFakeService',
+      );
+    }
+    return Future.value(_queuedMicrotasks);
+  }
+
+  @override
   Future<ClassList> getClassList(String isolateId) {
     return Future.value(_classList ?? ClassList(classes: []));
   }
@@ -539,4 +552,10 @@ class FakeVmServiceWrapper extends Fake implements VmServiceWrapper {
 
   @override
   Stream<Event> get onIsolateEvent => const Stream.empty();
+
+  @override
+  Stream<Event> get onTimerEvent => const Stream.empty();
+
+  @override
+  Stream<Event> get onTimerEventWithHistorySafe => const Stream.empty();
 }