+++ packages/devtools_app_shared/CHANGELOG.md
@@ -3,7 +3,13 @@ Copyright 2025 The Flutter Authors
 Use of this source code is governed by a BSD-style license that can be
 found in the LICENSE file or at https://developers.google.com/open-source/licenses/bsd.
 -->
-## 0.4.0 (not released)
+## 0.4.1 (not released)
+* Ignore expected exceptions from the `navigateToCode` extension method.
+* Add `RpcErrorExtension` extension with an `isServiceDisposedError` getter.
+
+## 0.4.0
+* Bump `dtd` dependency to `^4.0.0`.
+* Bump `devtools_shared` dependency to `^12.0.0`.
 * Add `DisposableController.init` method and `DisposableController.disposed`
 getter.
 * Move the `Disposable` class from the